[GLB-40] image-metadata 페이지 리팩토링 및 컴포넌트 분리 (#178)
* refactor: ImageMetadata의 비즈니스 로직을 커스텀 훅으로 추출

- 이미지 상태 관리 및 파일 처리 로직을 useImageMetadata로 이동
- 다이어리 저장 및 수정 로직을 useDiaryAction으로 이동

* refactor: 가독성 향상을 위해 ImageMetadata 하위 컴포넌트 분리

- 이미지 리스트 및 업로드 플레이스홀더를 ImageUploadSection으로 분리
- 도시 ID 오류 상태 UI를 CityErrorFallback으로 분리

* refactor: 추출한 훅과 하위 컴포넌트를 ImageMetadata에 적용

- 전체 코드량을 약 730줄에서 110줄로 축소
- 관심사 분리를 통해 컴포넌트 구조 개선

* refactor: addDiaryPhoto 타입 추론 개선 및 payload 빌더 로직 모듈화

* refactor: isCoordinateFormat 공통 좌표 검증 로직을 util 함수로 분리
